SQL refactor POC
