+++ datastore/datastoretest/mock_datastore.go
@@ -46,13 +46,13 @@ func (m *MockDatastore) HasItem(clientID string, ID string) (bool, error) {
 }
 
 // GetClientItemCount mocks calls to GetClientItemCount
-func (m *MockDatastore) GetClientItemCount(clientID string) (*datastore.ClientItemCounts, error) {
+func (m *MockDatastore) GetClientItemCount(clientID string) (*datastore.DynamoItemCounts, error) {
 	args := m.Called(clientID)
-	return &datastore.ClientItemCounts{ClientID: clientID, ID: clientID}, args.Error(1)
+	return &datastore.DynamoItemCounts{ClientID: clientID, ID: clientID}, args.Error(1)
 }
 
 // UpdateClientItemCount mocks calls to UpdateClientItemCount
-func (m *MockDatastore) UpdateClientItemCount(counts *datastore.ClientItemCounts, newNormalItemCount int, newHistoryItemCount int) error {
+func (m *MockDatastore) UpdateClientItemCount(counts *datastore.DynamoItemCounts, newNormalItemCount int, newHistoryItemCount int) error {
 	args := m.Called(counts, newNormalItemCount, newHistoryItemCount)
 	return args.Error(0)
 }
